refactor: add try/except/finally function for test_save_passwords_json

+++ tests/test_function.py
@@ -46,26 +46,28 @@ def test_save_passwords_csv():
 
 def test_save_passwords_json():
     """Test save_passwords_json function."""
-    test_passwords = [
-            {"service": "Google", "login": "user@gmail.com", "password": "P@ssw0rd!"},
-            {"service": "GitHub", "login": "username", "password": "Abc123!"}
-    ]
-
+    test_passwords = get_test_passwords()
     file_path = "test_passwords.json"
 
-    save_passwords_json(file_path, test_passwords)
+    try:
+        save_passwords_json(file_path, test_passwords)
 
-    assert os.path.exists(file_path), "File was not created!"
+        assert os.path.exists(file_path), "File was not created!"
 
-    with open(file_path, "r", encoding="utf-8") as f:
-        rows = json.load(f)
-        assert len(rows) == 2, "Wrong number of rows!"
-        assert rows[0]["service"] == "Google", "Wrong service name!"
+        with open(file_path, "r", encoding="utf-8") as f:
+            rows = json.load(f)
+            assert len(rows) == 2, "Wrong number of rows!"
+            assert rows[0]["service"] == "Google", "Wrong service name!"
 
-    print("✅ All tests for json format passed!")
+        print("✅ All tests for json format passed!")
 
-    os.remove(file_path)
-    print("🗑️  Test file removed.")
+    except Exception as e:
+        print(f"❌ Test failed:{e}")
+        raise
+    finally:
+        if os.path.exists(file_path):
+            os.remove(file_path)
+            print("🗑️ Test file removed.")
 
 
 def test_save_passwords_txt():
